feat: finishing UI
