+++ styles/LatestNews.module.css
@@ -1,3 +1,60 @@
 .last_news {
-    padding-left: 64px;
-  }
+  padding: 0 0px;
+  border-bottom: 1px solid black;
+}
+
+.last_news__header {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  gap: 48px;
+  border-bottom: 1px solid black;
+}
+
+.last_news__title {
+  text-transform: uppercase;
+  font-weight: 500;
+  line-height: 110%;
+  padding: 48px 0px;
+  margin: 0;
+}
+
+.last_news__news {
+  margin-top: 64px;
+  margin-bottom: 64px;
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  gap: 24px;
+}
+
+.default_button {
+  padding: 10px 16px;
+  width: 102px;
+  height: 35px;
+  background-color: var(--background-default);
+  border: 1px solid #000000;
+  border-radius: 6px;
+  color: #000000;
+  font-family: 'Roboto Mono';
+  font-style: normal;
+  font-weight: 400;
+  font-size: 15px;
+  line-height: 100%;
+  text-transform: uppercase;
+  letter-spacing: 1px;
+  cursor: pointer;
+}
+
+.inverted_button {
+  padding: 10px 16px;
+  width: 102px;
+  height: 35px;
+  background-color: transparent;
+  color: var(--text-default);
+  border-radius: 6px;
+  border: 1px solid var(--border-default);
+  letter-spacing: 1px;
+  text-transform: uppercase;
+  cursor: pointer;
+}
