+++ styles/OurValues.module.css
@@ -1,5 +1,5 @@
 .our_values {
-  padding: 0 68px;
+  padding: 0 0px;
 }
 
 .our_values__title {
